live-proxy/requirements.txt
websockets>=12.0
google-auth>=2.0
aiohttp>=3.9
certifi
requests
